requirements.txt
openai==1.1.1
tiktoken
python-dotenv
Pillow
ipython